+++ src/shared/Router.js
@@ -7,6 +7,7 @@ import SignUp from "../pages/user/SignUp";
 import ScheduleRegisterPage from "../pages/schedule/SceduleRegisterPage";
 import RestaurantsPage from "../pages/mainpage/RestaurantsPage";
 import AccommodationsPage from "../pages/mainpage/AccommodationsPage";
+import Registration from "../components/schedule/Resitration";
 
 const Router = () => {
   return (
@@ -17,7 +18,8 @@ const Router = () => {
         <Route path="/restaurants" element={<RestaurantsPage />} />
         <Route path="/accommodations" element={<AccommodationsPage />} />
         <Route path="/signup" element={<SignUp />} />
-        <Route path="/schedule/register" element={<ScheduleRegisterPage />} />
+        <Route path="/schedule" element={<ScheduleRegisterPage />} />
+        <Route path="/schedule/register" element={<Registration />} />
         <Route path="*" element={<NotFound />} />
       </Routes>
     </BrowserRouter>
